[nginx] Use variable in proxy_pass instead of upstream load balancer
These changes remove the upstream clause and the load balancer
capability of the reverse proxy. The issue with the previous
configuration was that without the commercial nginx version there is no
way for the load balancer to monitor containers for dynamic IP change.
The nginx caches the DNS but does not update its list of IP unless
forced to do so.
This means that in the case where a container goes off, then comes back
with a different IP then the reverse proxy is unable to detect that
change and will continue to return a 502 error.

It turns out that using variables as argument for the proxy_pass makes
nginx check the DNS for its IP (caches the result for 5min by default).
This solves the issue of tracking IP changes at the cost of not using
the load balancer.