feat(ansible): Make all of the host configuration go through ansible
This has been done by actually merging the contents of the repo
https://github.com/Nelyah/system-deploy and tailoring it specifically to
match the requirements of this server.

One major change is that the secret files are now encrypted using a GPG
key rather than using Ansible vault. This is to have a single "standard"
way of encrypting files throughout the repo.

This is still a first version and a few things remain to be done:
- Automating docker volumes creation
- Automating docker network creation
- Setup systemd backup units/timers
- Automating deployment of those backup units