ansible: Do not change the name of the backup by appending a date
This is causing a change in the arguments given to restic (because the
file is not the same). This, in turn, causes restic to RE-READ all of
the files instead of only looking at those which changed.

I don't know why Restic has this behaviour, but at list this fixes it.