caddy: Allow tailscale network as local network
Note: This requires that the tailscale network has a DNS running that is
able to resolve the domain name nelyah.eu (or the domain name that I want to use).

Currently I have something set up for that with dnsmasq manually, will
probably document it in Obsidian, I don't think that's something I want
to automate.